requirements.txt
absl-py==0.15.0
argcomplete==1.12.3
argon2-cffi==21.1.0
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cycler==0.11.0
Cython==0.29.24
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
Django==3.2.8
djangorestframework==3.12.4
entrypoints==0.3
google-auth==2.3.2
google-auth-oauthlib==0.4.6
grpcio==1.41.1
idna==3.3
imageio==2.9.0
importlib-metadata==4.8.1
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
jsonschema==4.1.2
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.2.2
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.5
numpy==1.21.3
oauthlib==3.1.1
onnx==1.10.2
opencv-python==4.3.0.38
packaging==21.2
pandocfilters==1.5.0
parso==0.8.2
pathlib==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.21
protobuf==3.19.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
Pygments==2.10.0
pyparsing
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.3
scipy==1.7.1
Send2Trash==1.8.0
six==1.16.0
sqlparse==0.4.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminado==0.12.1
testpath==0.5.0
thop==0.0.31.post2005241907
tifffile==2021.10.12
torch==1.10.0
torchvision==0.11.1
tornado==6.1
tqdm==4.47.0
traitlets==5.1.1
typing-extensions==3.10.0.2
urllib3==1.26.7
utils==1.0.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
zipp==3.6.0
mean-average-precision==2021.4.26.0
